Move to "devices" instead of inputs/outputs
This was a bad metaphor - some devices function as both inputs and outputs (ex. PowerMate has input rotation/press, but output LED light)

They're all just devices, they can talk to each other however they want - no need to be arbitrary with what kind of device it is
