+++ knob-ts/src/devices/powermate.ts
